Change delimiting scheme.
Instead of '>' and '$' prefixes and a ';' to terminate,
we just use '.' and ',' to terminate.

"X." means parse and run X on the unconsumed input Y.
It should return a string, which we print.

"X," means parse and run X on the unconsumed input Y.
It should return a string X' and possibly leave some of Y
unconsumed, which we denote Y'. Then parse and run X' on Y'.