Experimenting with VM.
Number the basic combinators consecutively from 0.
Should make jump tables more attractive.

After trial and error with GCC 9.2.0 and Clang 8.0.0 on 11-queens:

For Clang, we put the stack in the unused heap, and short-circuit application
nodes.

For GCC, we catch the application in the default case of a switch statement.

I got confused because I failed to realize some of my combinator rewrite
"optimizations" can create a lot of garbage. They happened to speed up
11-queens, but other programs slowed down. After removing them, it seems
the VM is worse than before. I'm committing the VM changes for posterity.