Fix top-level type annotation bug.
Suppose a function has inferred type:

  Eq a => a -> Bool

And we annotate it with the type:

  Char -> Bool

The old code searched for suitable instances for the inferred type first, then
it matched the inferred type against the annotated type to augment the
resulting AST to specialize. In the above example it would add a lambda and an
application to the `Eq Char` dictionary. However, it neglected to do this for
any recursive calls.

The new code is simpler: first we match the inferred type against the annotated
type, and substitute the types in  the `Proof` nodes accordingly. Then we just
search for suitable instances to produce the desired AST.

There's still more to do. For example, type contexts of annotations are ignored
completely!