wasm: remove generation of `go` from `main`.
It assumed `main()` is only invoked once, which is may be false for wasm.
After a garbage collection, the heap address hardcoded in the generated
`go()` function is almost certainly wrong.

The programmer should instead declare a foreign export for `main`.
Our code records its heap address in the `root` table, so it will be
updating during garbage collection.