Spin off type information from `Link`.
This is only used during type inference, so the type field was often
`undefined`, or unnecessarily set in the case of foreign imports.

We bundle a type with a `Link` because we already looked it up when resolving
symbols from other module. This saves a lookup as well as passing around
symbol tables of other modules at the cost of an awkward `E (Basic "@")`
node.