"Slide Y" optimization.
Detect when all recursive calls start with the same arguments and
place a Y combinator so that each argument is only bound once
rather than on each recursive call.

I was motivated to do this because of the example:

  fibs = 0 : 1 : zipWith (+) fibs (tail fibs)

This is polymorphic with `precisely`, and passing the `Ring`
dictionary every call gives it exponential complexity.