REPL fixes.
Allowing definitions to use older definitions is trickier than
I thought. The simplest solution seems to be to import a special
module holding definitions entered in the REPL, which we call ">"
to avoid any possible collisions, and modify the ambiguity check
to ignore this special case so functions can be redefined.