Remove dead code.
The parser adds default typeclass methods to `topDefs`. But then
`genDefaultMethods` looks each of them up in `typedAsts` instead.
Since the lookup fails, a `fail#` definition is generated.

During type inference, the `topDefs` version overwrites the `fail#`
definition so long as it is well-typed. At this point, there is no
check that its type matches the declaration in the typeclass.

For each instance, the type of each default method is matched
against the typeclass declaration.

We save steps by removing `genDefaultMethod`. In later compilers,
we generate `fail#` in `inferTypeclasses` for undefined methods,
a feature which we have yet to use.