atp: replace Haste with our compiler.
Found a nasty bug while doing so: the `optiApp` function is
dangerous because of variable capture. It's meant to simplify
redundant joins and perhaps shrink dictionary code, but it clashed
badly with `triangulate`, which performs one-off substitutions
reusing many variables.

I guess we never noticed because I never tried enough mutually
recursve expressions in a let block.

For now, we spare `optiApp`, and change `triangulate` to use `fill`
to rewrite strongly connected components in let blocks.

precisely: we call `fill` after inserting `Proof` nodes so we
had to extend it to handle this case.