stitch: menu generation.
Encountered another AsciiDoctor migration issue. The rules for
substitutions in passthrough blocks seem different. Instead of
trying to figure it out, we take advantage of AsciiDoctor's
concise markup for `id` and `class` attributes.

AsciiDoctor treats `xs ++ ys` differently to AsciiDoc, which
I worked aorund with by changing an example.

I've also lost syntax highlighting, but fixing this can wait.

Forgot to mention: as I was revamping the HTML anyway, I switched
the navbar arrows from Unicode 9664 and 9654 to 9204 and 9205
whose smaller size somehow caused the article heading to be
bumped to the right. I worked around it by padding the top of
the heading.