Experiment with `global` feature.
It may be safer than an `unsaferPerformIO newIORef`, because
modifying `global` has no effect on the heap, unlike `writeIORef`.
The latter leads to bugs when the heap address of the underlying
`IORef` is shared, which is difficult to prevent in all cases
because due to `unsafePerformIO`.

Support heap indexes in foreign functions. This simplifies
`expy` and `vmdump`.