Clean up more early compilers.
Some versions of `rank` would return the input symbol if no corresponding
address could be found, a strange fallback behaviour that some compilers
depended on for the cons function `(:)`. Now, it is more principled: a missing
symbol is considered an error. Also, instead of a special `primTab` lookup
table, we preprend built-in definitions to the program.